+++ CoinTree/Models/CoinPriceDetails.cs
@@ -7,10 +7,10 @@ namespace CoinTree.Models
     {
         public string Sell { get; set; }
         public string Buy { get; set; }
-        public decimal Ask { get; set; }
-        public decimal Bid { get; set; }
-        public decimal Rate { get; set; }
-        public decimal SpotRate { get; set; }
+        public double Ask { get; set; }
+        public double Bid { get; set; }
+        public double Rate { get; set; }
+        public double SpotRate { get; set; }
         public string Market { get; set; }
         public DateTime Timestamp { get; set; }
         public string RateType { get; set; }
